ml_pipeline/requirements.txt
# ML Pipeline Dependencies
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
optuna>=3.4.0
onnx>=1.15.0
onnxruntime>=1.16.0
matplotlib>=3.7.0
seaborn>=0.13.0
pyyaml>=6.0
tqdm>=4.66.0
Pillow>=10.0.0
albumentations>=1.3.0
tensorboard>=2.15.0
netcal>=1.3.0
